Factor out `ChainSyncEventTrace` & `LabelPeerTrace` to their own modules
