+++ cardano-node/cardano-node.cabal
@@ -45,10 +45,12 @@ library
 
   other-modules:       Paths_cardano_node
                        Cardano.Tracing.ToObjectOrphans.ChainDBTrace
+                       Cardano.Tracing.ToObjectOrphans.ChainSyncEventTrace
                        Cardano.Tracing.ToObjectOrphans.DnsTrace
                        Cardano.Tracing.ToObjectOrphans.ErrorPolicyTrace
                        Cardano.Tracing.ToObjectOrphans.MuxTrace
                        Cardano.Tracing.ToObjectOrphans.SubscriptionTrace
+                       Cardano.Tracing.ToObjectOrphans.LabelPeerTrace
                        Cardano.Tracing.ToObjectOrphans
                        Cardano.Tracing.MicroBenchmarking
