+++ cardano-node/src/Cardano/Tracing/Tracers.hs
@@ -68,8 +68,10 @@ import qualified Ouroboros.Storage.ChainDB as ChainDB
 import           Cardano.Config.Protocol (TraceConstraints)
 import           Cardano.Config.Types
 import           Cardano.Tracing.ToObjectOrphans.ChainDBTrace ()
+import           Cardano.Tracing.ToObjectOrphans.ChainSyncEventTrace ()
 import           Cardano.Tracing.ToObjectOrphans.DnsTrace ()
 import           Cardano.Tracing.ToObjectOrphans.ErrorPolicyTrace ()
+import           Cardano.Tracing.ToObjectOrphans.LabelPeerTrace ()
 import           Cardano.Tracing.ToObjectOrphans.MuxTrace ()
 import           Cardano.Tracing.ToObjectOrphans.SubscriptionTrace ()
 import           Cardano.Tracing.ToObjectOrphans
